Move TimeZone and TimeZoneVariant outside of provider module
